Prevent exception when language service not enabled
Fixes #8500

In some cases the language service host environment may signal that language services should not be enabled. Command line mode /build is an example of this.

Users of the language service receive exceptions when attempting to use its APIs when it is disabled.

This change allows callers to validate that the language service is enabled in order to avoid such exceptions.

#8502 tracks using project capabilities instead of this manual check. As part of that work, we would revert this change, instead updating the project capabilities on `LanguageServiceErrorListProvider` itself (or just making the `LanguageServices` capability dynamic).