Replace custom browse object base class
In the past, we had to add our own base class for dependency tree browse objects in order to customise the component name displayed in the Properties pane. The VS shell provided a similar base class, but the `GetComponentName` wasn't virtual, so we couldn't customise it.

I fixed that in https://dev.azure.com/devdiv/DevDiv/_git/VS/pullrequest/248337

This change removes our custom implementation, replacing it with the updated class from the VS shell.