No fault when insufficient data to create workspace
There may be projects that do not provide enough information for us to create the workspace. An example of this is `.ilproj`. We should not fault the project in such cases by throwing. Instead, we dispose the workspace which tears down its dataflow and cleans up gracefully.

This restores the behaviour we had prior to the changes made to the language service code to improve performance of configuration switches.