Move NetCoreRuntimeVersionsRegistryReader.cs
The code in this type is only used by the `MissingSetupComponentRegistrationService`, and is effectively an implementation detail of that type. Here we move the file from the Microsoft.VisualStudio.ProjectSystem.Managed project to Microsoft.VisualStudio.ProjectSystem.Managed.VS so the definition is closer to where it is used.

A future commit will update the namespace to match the folder.