Remove reference to IProjectAsyncLoadDashboard
The `IProjectAsyncLoadDashboard` interface comes from the VS layer of CPS, but we have a reference to it in a doc comment in Microsoft.ProjectSystem.VisualStudio.Managed.csproj. This removes the reference in order to enforce proper layering.

While I was at it, I updated the affected file to use a file-scoped namespace.