Update FlushUnshippedAPIFiles.ps1
The PublicAPI.Shipped.txt file in each of our projects documents that the public API at the time of the last GA release, and an analyzer warns us when it is out of sync with the project's code. Public API changes are often breaking changes, which are of course to be avoided if possible (or at least carefully managed). With the analyzer helping us keep the PublicAPI.Shipped.txt file in sync with the code, we can easily see changes to the API in a code review.

There is also a PublicAPI.Unshipped.txt file, which is meant to hold _new_ public APIs that have not yet been released. Modification to these APIs are less problematic, as they haven't been released in a GA version of the product yet and in theory no one has taken a dependency on them. In theory, when we release a new version of the project we move all the items in the "unshipped" file to the "shipped" file.

In practice we have often forgotten to do this step with the result that items build up in the "unshipped" file over multiple releases, making it much less useful. Also, internal teams tend to take dependencies on our new APIs before GA, so removing or altering a theoretically "unshipped" API is still very problematic for us. So we don't really use PublicAPI.Unshipped.txt, and we have a script--FlushUnshippedAPIFiles.ps1--that moves any items in the unshipped file to the shipped file on every build.

This change updates the script to handle multiple pairs of Shipped/Unshipped files within a single project, and to drop the assumptions on where those files will be located. Instead it assumes that every PublicAPI.Unshipped.txt file will have the corresponding PublicAPI.Shipped.txt right next to it. It also uses "Write-Host" to call out when it is copying content from one file to another; this message will end up in the build log which can aid in diagnosing problems.