Only run FlushUnshippedAPI on the outer build
The FlushUnshippedAPIFiles.ps1 script will update *all* public API files in a project, not just the ones for a particular target framework. As such we should only run it on the "outer" build; this prevents it from running more than it needs to. It also prevents multiple invocations of the script from running at the same time which could lead to file locking issues as they try to update the same files.