Specify TargetFramework in each project
Now that we have a couple of projects that use multi-targeting it no longer makes sense to set `<TargetFramework>` once in a Directory.Build.props file and have that flow down to the projects. Here we remove that property from the shared .props file and instead set the TargetFramework or TargetFrameworks in each individual project.