Update CPS packages
The CPS changes to build against .NET 6.0 have merged into their main branch, and this PR picks up the packages produced from that branch.

This version of their packages also drops Microsoft.IO dependencies from the .NET 6.0 build, which means our .NET 6.0 build no longer gets a transitive dependency. This caused a build error in the `WindowsFileExplorer` type which utilizes `Microsoft.IO.Path`.

The `WindowsFileExplorer` type explicitly opens files in the "explorer.exe" process which is of course only available on Windows. Since MS.VS.PS.Managed now needs to work across operating systems the fix here is to move `WindowsFileExplorer` up to the VS layer.

The `IFileExplorer` interface it implements is defined in the MS.VS.PS.Managed layer but not used there, so for the moment it is OK that it will have no implementation in VS Code scenarios.