Add analyser to catch multiple enumeration
This is a pattern that comes up in PR review fairly often. It's easy to inadvertently enumerate a lazy sequence more than once in code, which can have undesirable performance characteristics, or produce invalid behaviour.

This change enables an analyzer that catches such issues, and fixes the existing occurrences of it in our code base.