Move TargetFrameworkProvider
The `ITargetFrameworkProvider` interface is defined in Microsoft.VisualStudio.ProjectSystem.Managed.csproj, and all usages of that type are in that project as well. However, the _implementation_ of the type is in Microsoft.VisualStudio.ProjectSystem.Managed.VS.csproj. There's no good reason for this split; this change moves the implementation down a layer to live next to the definition.