Move language service support
The https://github.com/dotnet/roslyn/pull/65592 the Roslyn team defined a new brokered service for communicating information from the project system to the language service. In effect, the `IWorkspaceProjectContext` and `IWorkspaceProjectContextFactory` types are being replaced by the very similar `IWorkspaceProject` and `IWorkspaceProjectFactoryService` types, but it isn't happening all at once. Instead, the VS Code-specific version of the project system will adopt the new interfaces first (since the old ones don't work at all in a system where the language service and project system are in different processes) and at a later date the VS for Windows-specific project system will move over, once we've worked out all the bugs. This should prevent regressions in VS for Windows while we get VS Code up and running.

However, this means that any code depending on `IWorkspaceProjectContext` and `IWorkspaceProjectContextFactory` needs to move out of our host-agnostic layer (Microsoft.VisualStudio.ProjectSystem.Managed) and into our VS-for-Windows-specific layer (Microsoft.VisualStudio.ProjectSystem.Managed.VS). This commit does the minimum necessary to accomplish that:

1. Any files that actually call the old interfaces are moved as-is to the VS layer.
2. Any further files that depend on those files in 1 are moved as well.
3. Files that refer to the old interfaces only in comments or doc comments have been updated to use phrases like "language service project" instead but have otherwise been left in place.
4. The dependency on the Microsoft.VisualStudio.LanguageServices package has moved from the host-agnostic layer to the VS-specific layer.
5. The `DotNetImportsEnumProvider` has also moved to the VS layer, as it depends on the `VisualStudioWorkspace` type. The host-agnostic layer no longer references the assembly with that type.
6. Unit test files have also been moved around to reflect the new layering.

There are a few further dependencies on the Roslyn `Workspace` type in Microsoft.VisualStudio.ProjectSystem.Managed; those will be moved in a future change as we can't count on being able to MEF-import a `Workspace` in VS Code scenarios; it will generally exist only in the language service process.