Optimize copy item checking when build acceleration is disabled
Build Acceleration is a new feature that requires opt in.

When adding that feature, we started checking a larger number of potential inputs to projects. This included files like the output assemblies of referenced projects, along with their debug symbols and documentation files.

We need to check such files for the Build Acceleration feature, but when it is disabled there is no need to check them and doing so creates additional overhead that makes the FUTDC slower.

This change adds new metadata to those items, `BuildAccelerationOnly`, so the FUTDC can decide whether to check them or not.

That change filters all items from many projects during builds, leading to confusing messages being logged. Quite a bit of the diff here relates to moving logging code around to produce better quality log messages. In doing so, a more robust approach to controlling nesting of log messages was introduced. The diff is easier to read if white space changes are ignored.

We also remove the duplicate item checking that we saw previously as I never saw this being hit in practice, and tracking these items has a cost that I no longer believe is worth it.

Unit tests were extended to include items both with and without `BuildAccelerationOnly` metadata to validate behaviour. Note that the logging changes also created some diff in the unit tests of unrelated methods, however those changes mean that the tests now reflect the output that is actually seen when running the FUTDC in VS.