Identify FUTDC time spent waiting for project data
In current telemetry we report a single duration for the FUTDC. In practice, there are two interesting components to this number:

1. Time spent waiting for project data. The subscription object calls GetLatestVersionAsync for project data, which can take a while. We also acquire a lock and do some other work during this time.
2. Time spent executing the check itself, iterating over files and so forth.

With this change, we add another measure to our telemetry that captures only the wait time. This will allow us to subtract this from the previously reported number (which we continue to report unchanged) so we can more easily identify wait time versus the check's actual execution time.