Remove use of obsolete type
As of .NET 6, the `SHA256CryptoServiceProvider` type (among others) is obsolete. Instead, the guidance is to call `SHA256.Create()` to create the provider.

Note that since we're hashing a single value we could have instead used one of the `SHA256.HashData(...)` overloads and avoided the allocation of a `SHA256` instance altogether. However, `HashData` is not available on net472, while the approach we adopt here will work for both net472 and net6.0.

See https://learn.microsoft.com/en-us/dotnet/fundamentals/syslib-diagnostics/syslib0021 for more information.