Use the correct IAsyncDisposable by default
There are two definitions of `IAsyncDisposable`:

1. `System.IAsyncDisposable`
2. `Microsoft.VisualStudio.Threading.IAsyncDisposable`

We only want to use the first. This commit adds an alias for the symbol so that it's hard to accidentally use the wrong version. If we ever do need to use the other definition, we can still do so by being explicit about it.