Up-to-date check doesn't require ordered snapshot data
CPS uses immutable dictionaries to model snapshot data. The default implementations of these reorder their items (by hash code, which is essentially random).

Snapshot data that comes from CPS builds is backed by interface `IDataWithOriginalSource<>` which allows access to data in the original order.

It turns out this interface is not available on data sourced from evaluation.

Because the FUTDC uses the `JointRuleSource` it receives a mix of both evaluation and build data. Therefore in a given snapshot we cannot guarantee item ordering.

This change attempts to use ordered data when available, but falls back to the random order when it needs to. The order of items here is not important. The original reason for adding ordered data here was to make unit tests stable across runs, where we check the logged output of data. Those tests will continue to run, as we stage mock data that always implements this interface.