Add custom dictionary for spell checker
VS now supports adding a custom dictionary for spell checking. This allows us to stop seeing so many squiggles in code due to use of non-English words that are perfectly valid in the context of our software.

Note that, for now at least, this file must be UTF-16 with a BOM to work correctly. Git treats this file as binary, so merging it might be tricky.