Fix a property name
On the VB "Build" property page we have a property named "Output path". Note that is not the text displayed to the user, but the actual ID of the property on the page. When the user attempts to set this property they'll get exceptions from MSBuild complaining that MSBuild property names cannot include spaces; this is because the property name on the page is used as-is for the property name in the underlying MSBuild file.

Here we fix the property name on the .xaml file, and add a unit test that scans all our .xaml files and verify that none of the property names contain spaces.