Reduce exception message size
Our fault handling logic wraps exceptions raised on Dataflow blocks and reports them via `IProjectFaultHandlerService`.

The previous implementation was constructing that wrapper exception such that its `Message` contained the full `ToString()` of the inner exception. That inner string can be enormous for deep stacks, or deeply nested chains of exceptions (or both).

Lifeng observed instances of this exception string crowding the Large Object Heap, and this fix should address that by simplifying the message of the wrapper exception.