FUTDC excludes non-modifiable input items
Items in certain locations on disk are considered non-modifiable, and therefore are not inspected for changes during the fast up-to-date check.

Previous work in this area filtered various references, such as framework assemblies, using our `ProjectFileClassifier`.

This change extends the behaviour to input items, primarily to exclude items added via NuGet packages such as `Compile` and `AdditionalFiles` items.