Move NuGet.VisualStudio dependency to VS layer
Currently the code to collect the information needed for NuGet restore operations lives in the Microsoft.VisualStudio.ProjectSystem.Managed layer, though the actual call to the VS NuGet APIs (specifically, `IVsSolutionRestoreService3.NominateProjectAsync`) only occurs in the Microsoft.VisualStudio.ProjectSystem.Managed.VS layer. Even though the *.Managed layer doesn't use that API, it does still have a dependency on the NuGet.VisualStudio NuGet package because the data structures it builds up implement a number of NuGet interfaces (`IVsReferenceProperties`, `IVsProjectRestoreInfo2`, `IVsReferenceItem`, etc.).

The *.Managed layer sometimes makes use of these interfaces, but it doesn't _need_ to--it already knows what the concrete type is. We also have a number of collection types (`ProjectProperties`, `ReferenceItems`, etc.) that exist only to implement a NuGet collection interface (such as `IVsProjectProperties` or `IVsReferenceItems`) that allow access by either index or name. We don't need this functionality at the *.Managed layer, either, and can quite easily use an `ImmutableList<T>` instead of a custom collection type.

This commit strips the NuGet interfaces from the data model in the *.Managed layer, and eliminates the unnecessary collection types. This in turn allows us to drop the reference to the NuGet.VisualStudio package.

In the *.Managed.VS layer we add a number of wrapper types that exist to adapt the data model to the NuGet-specific interfaces. Where possible, the creation of the wrappers is done lazily; if NuGet never queries a particular property of the data model, no wrapper will be created.

The most interesting part of this change is to the `ImmutablePropertyCollection` which is what allows us to create collections that support accesing items by index or name. It will take as inputs types of the data model from the *.Managed layer, but it needs to offer to the consumer (NuGet) objects with the expected NuGet interface. It now takes a transform function, allowing derived types to specify how to create an "output" object from an "input".

A future change will move the `ImmutablePropertyCollection` type up to the *.Managed.VS layer as it is no longer used in *.Managed.

There is more that could be done to simplify the data model. The `ProjectProperty` and `ReferenceProperty` types are identical and could be replaced by a `KeyValuePair<string, string>`, and all the data is originally in dictionaries in `IProjectRuleSnapshot` instances. We could pass those dictionaries around instead of converting them to a NuGet-specific object model which would reduce the number of types and avoid some object allocations, but it isn't necessary and would create too much churn at this point. We can come back to that at a later date if we wish.