Create a non-model user notification service
The `PackageRestoreDataSource` has a dependency on the VS info bar via the `IInfoBarService` which is of course defined in our VS layer. I'd like to move `PackageRestoreDataSource` down to MS.VS.PS.Managed, so it needs to be updated to use a similar service from that layer.

As we currently don't have anything suitable at that layer, this change introduces the `INonModalNotificationService`. This service provides a way of showing simple messages, warnings, and errors that does not block the user (e.g. it does not pop up a modal dialog). The VS layer implementation of this delegates to the existin `IInfoBarService`.