Rename *Wrapper types to Vs*
Rename the `*Wrapper` types used to implement the various `IVs*` interfaces defined by NuGet to `Vs*` instead. For example, `ProjectPropertyWrapper` implements `IVsProjectProperty` and here is renamed to `VsProjectProperty`. This emphasizes that the types are used to communicate information to a different part of VS rather than used internally.

The files will be renamed in a subsequent commit.