Prefer ImmutableArray over ImmutableList
`ImmutableArray` is the thinnest possible wrapper (just a struct) over a straight array, while `ImmutableList` is an object that will wrap at least one other object and possibly as much as one object per list element. `ImmutableList` is a better choice when we expect significant transformations to the list (lots of adding and removing items) as it may be able to reuse the underlying objects; `ImmutableArray` can't do that and will always end up allocating a new array of the necessary size.

In the case of the NuGet restore information, however, we basically create it once and transform it relatively rarely if at all. In this case `ImmutableArray` is the better option as it will significantly reduce the number of object allocations.