Import IProjectFaultHandlerService
The `NuGetRestoreService` uses the `IProjectFaultHandlerService` to run some "fire and forget" work. Since the work is "forgotten" our unit tests can't wait for it to complete, which leads to a couple of flaky (and currently skipped) tests.

Right now we're acquiring the service from `UnconfiguredProject.Services`, but by importing it directly in the constructor the unit tests can provide a mock that captures the otherwise "forgotten" `Task` and then await it.